pipeline tetsing loacally

+++ Jenkinsfile
@@ -2,8 +2,9 @@ pipeline {
     agent any
 
     environment {
-        DOCKER_HUB = credentials('dockerhub-username')
+        DOCKER_HUB = "kunalmane"  // replace with your Docker Hub username
         IMAGE_TAG = "v${BUILD_NUMBER}"
+        DEPLOY_LOCAL = "true"
     }
 
     stages {
@@ -42,21 +43,30 @@ pipeline {
         }
 
         stage('Deploy Containers') {
+            when {
+                expression { return env.DEPLOY_LOCAL == "true" }
+            }
             steps {
                 sh 'docker-compose down && docker-compose up -d'
             }
         }
 
-        stage('Push Images') {
+        stage('Smoke Test') {
             steps {
-                withCredentials([usernamePassword(credentialsId: 'dockerhub-creds', usernameVariable: 'DOCKER_USER', passwordVariable: 'DOCKER_PASS')]) {
-                    sh '''
-                    echo $DOCKER_PASS | docker login -u $DOCKER_USER --password-stdin
-                    docker push $DOCKER_HUB/frontend:$IMAGE_TAG
-                    docker push $DOCKER_HUB/backend:$IMAGE_TAG
-                    '''
-                }
+                sh 'curl -f http://localhost:8090 || exit 1'
             }
         }
+
+        // stage('Push Images') {
+        //     steps {
+        //         withCredentials([usernamePassword(credentialsId: 'dockerhub-creds', usernameVariable: 'DOCKER_USER', passwordVariable: 'DOCKER_PASS')]) {
+        //             sh '''
+        //             echo $DOCKER_PASS | docker login -u $DOCKER_USER --password-stdin
+        //             docker push $DOCKER_HUB/frontend:$IMAGE_TAG
+        //             docker push $DOCKER_HUB/backend:$IMAGE_TAG
+        //             '''
+        //         }
+        //     }
+        // }
     }
 }
